Move things into lib
